Allow user to confirm load before start, re #8435

+++ arches/app/etl_modules/branch_csv_importer.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+import glob
 import json
 import logging
 import math
@@ -173,32 +174,28 @@ class BranchCsvImporter:
                     pass
         return {"name": worksheet.title, "rows": row_count}
 
-    def stage_excel_file(self, dir, summary):
-        for (dirpath, dirnames, filenames) in os.walk(dir):
-            for filename in filenames:
-                if filename.endswith("xlsx"):
-                    summary["files"][filename]["worksheets"] = []
-                    workbook = load_workbook(filename=os.path.join(dirpath, filename))
-                    try:
-                        graphid = workbook.get_sheet_by_name("metadata")["B1"].value
-                    except KeyError:
-                        with connection.cursor() as cursor:
-                            cursor.execute(
-                                """UPDATE load_event SET status = %s WHERE loadid = %s""",
-                                ("failed", self.loadid),
-                            )
-                        raise ValueError("A graphid is not available in the metadata worksheet")
-                    nodegroup_lookup, nodes = self.get_graph_tree(graphid)
-                    node_lookup = self.get_node_lookup(nodes)
-                    with connection.cursor() as cursor:
-                        cursor.execute(
-                            """UPDATE load_event SET load_details = %s WHERE loadid = %s""",
-                            (json.dumps(summary), self.loadid),
-                        )
-                        for worksheet in workbook.worksheets:
-                            if worksheet.title.lower() != "metadata":
-                                details = self.process_worksheet(worksheet, cursor, node_lookup, nodegroup_lookup)
-                                summary["files"][filename]["worksheets"].append(details)
+    def stage_excel_file(self, file, summary, cursor):
+        if file.endswith("xlsx"):
+            summary["files"][file]["worksheets"] = []
+            workbook = load_workbook(filename=os.path.join(self.temp_dir, file))
+            try:
+                graphid = workbook.get_sheet_by_name("metadata")["B1"].value
+            except KeyError:
+                cursor.execute(
+                    """UPDATE load_event SET status = %s WHERE loadid = %s""",
+                    ("failed", self.loadid),
+                )
+                raise ValueError("A graphid is not available in the metadata worksheet")
+            nodegroup_lookup, nodes = self.get_graph_tree(graphid)
+            node_lookup = self.get_node_lookup(nodes)
+            for worksheet in workbook.worksheets:
+                if worksheet.title.lower() != "metadata":
+                    details = self.process_worksheet(worksheet, cursor, node_lookup, nodegroup_lookup)
+                    summary["files"][file]["worksheets"].append(details)
+            cursor.execute(
+                """UPDATE load_event SET load_details = %s WHERE loadid = %s""",
+                (json.dumps(summary), self.loadid),
+            )
 
     def get_node_lookup(self, nodes):
         lookup = {}
@@ -206,19 +203,14 @@ class BranchCsvImporter:
             lookup[node.alias] = {"nodeid": str(node.nodeid), "datatype": node.datatype, "config": node.config}
         return lookup
 
-    def start(self, request):
-        self.loadid = request.POST.get("load_id")
-        with connection.cursor() as cursor:
-            cursor.execute(
-                """INSERT INTO load_event (loadid, etl_module_id, complete, status, load_start_time, user_id) VALUES (%s, %s, %s, %s, %s, %s)""",
-                (self.loadid, self.moduleid, False, "running", datetime.now(), self.userid),
-            )
-        return {"success": True, "data": "event created"}
-
     def read(self, request):
         self.loadid = request.POST.get("load_id")
         content = request.FILES["file"]
         self.temp_dir = os.path.join(settings.APP_ROOT, "tmp", self.loadid)
+        try:
+            shutil.rmtree(self.temp_dir) #Remove dir if it already exists. os.mkdir won't overwrite
+        except (FileNotFoundError):
+            pass
         os.mkdir(self.temp_dir, 0o770)
         result = {"summary": {"name": content.name, "size": self.filesize_format(content.size), "files": {}}}
         with zipfile.ZipFile(content, "r") as zip_ref:
@@ -228,17 +220,34 @@ class BranchCsvImporter:
                     if not file.is_dir():
                         result["summary"]["files"][file.filename] = {"size": (self.filesize_format(file.file_size))}
                     zip_ref.extract(file, self.temp_dir)
-        self.stage_excel_file(self.temp_dir, result["summary"])
-        result["validation"] = self.validate(request)
-        if len(result["validation"]["data"]) == 0:
-            self.write(request)
-        else:
+        return {"success": result, "data": result}
+
+    def start(self, request):
+        self.loadid = request.POST.get("load_id")
+        self.temp_dir = os.path.join(settings.APP_ROOT, "tmp", self.loadid)
+        self.file_details = request.POST.get("load_details", None)
+        result = {}
+        if self.file_details:
+            details = json.loads(self.file_details)
+            files = details['result']['summary']['files']
+            summary = details['result']['summary']
             with connection.cursor() as cursor:
                 cursor.execute(
-                    """UPDATE load_event SET status = %s WHERE loadid = %s""",
-                    ("failed", self.loadid),
+                    """INSERT INTO load_event (loadid, etl_module_id, complete, status, load_start_time, user_id) VALUES (%s, %s, %s, %s, %s, %s)""",
+                    (self.loadid, self.moduleid, False, "running", datetime.now(), self.userid),
                 )
-        shutil.rmtree(self.temp_dir)
+                for file in files.keys():
+                    self.stage_excel_file(file, summary, cursor)
+                result["validation"] = self.validate(request)
+                if len(result["validation"]["data"]) == 0:
+                    self.write(request)
+                else:
+                    cursor.execute(
+                        """UPDATE load_event SET status = %s WHERE loadid = %s""",
+                        ("failed", self.loadid),
+                    )
+            shutil.rmtree(self.temp_dir)
+            result['summary'] = summary
         return {"success": result["validation"]["success"], "data": result}
 
     def validate(self, request):
